{EQEmu PR 3427} [Cleanup/Feature] Add support for bots to #showstats/#mystats
# Notes
- Cleans up a lot of the display in the popup window and uses tables instead of `.` and `&nbsp;`.
- Optionally displays a lot more data versus always displaying it, allowing us more room for things within the window.
- Converts `Client::SendStatsWindow(client, use_window)` to a Mob method.
- Converts `Mob::IsPlayerClass(class_id)` to a utility method.
- Converts `Mob::IsPlayerRace(race_id)` to a utility method.
- Adds `GetPlayerClassAbbreviation(class_id)` as a utility method.
- Adds `GetPlayerRaceAbbreviation(race_id)` as a utility method.
- Cleans up `#mystats` logic.
- Cleans up `#showstats` logic.

# Images
## Bot
![image](https://github.com/EQEmu/Server/assets/89047260/0bf103f9-9e6d-4782-b2aa-a927f720d04b)
## Client
![image](https://github.com/EQEmu/Server/assets/89047260/92e6393e-d7fd-497c-acb9-a20c89a185ce)

+++ zone/client_mods.cpp
@@ -577,7 +577,7 @@ int64 Client::CalcBaseMana()
 int64 Client::CalcBaseManaRegen()
 {
 	uint8 clevel = GetLevel();
-	int32 regen = 0;
+	int64 regen = 0;
 	if (IsSitting() || (GetHorseId() != 0)) {
 		if (HasSkill(EQ::skills::SkillMeditate)) {
 			regen = (((GetSkill(EQ::skills::SkillMeditate) / 10) + (clevel - (clevel / 4))) / 4) + 4;
